multiple VEU IPs support
This patch changes code of the VEU UIO registration and
some of APIs for multile VEUs support.

shveu_probe(), that is a new API, may search and register all
VEU IPs, and this should be invoked once before shveu_*() used.
The shveu_open() and shveu_close() requires an id number that
specifies which VEU used.